loop-agents/pyproject.toml
[project]
name = "loop-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.37.37",
    "google-adk==0.2.0",
    "langchain-aws==0.2.21",
    "litellm==1.67.0",
    "python-dotenv==1.1.0",
]
